requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
annotated-types=0.7.0=pypi_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2024.8.30=hf0a4a13_0
cachetools=5.5.0=pypi_0
certifi=2024.8.30=pypi_0
charset-normalizer=3.3.2=pypi_0
google-ai-generativelanguage=0.6.6=pypi_0
google-api-core=2.19.2=pypi_0
google-api-python-client=2.143.0=pypi_0
google-auth=2.34.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-generativeai=0.7.2=pypi_0
googleapis-common-protos=1.65.0=pypi_0
grpcio=1.66.1=pypi_0
grpcio-status=1.62.3=pypi_0
httplib2=0.22.0=pypi_0
idna=3.8=pypi_0
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.46.1=hc14010f_0
libzlib=1.3.1=hfb2fe0b_1
markdown-it-py=3.0.0=pypi_0
mdurl=0.1.2=pypi_0
ncurses=6.5=h7bae524_1
openssl=3.3.2=h8359307_0
pip=24.2=pyh8b19718_1
prompt-toolkit=3.0.47=pypi_0
proto-plus=1.24.0=pypi_0
protobuf=4.25.4=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.1.4=pypi_0
python=3.12.5=h30c5eda_0_cpython
python-dotenv=1.0.1=pypi_0
readline=8.2=h92ec313_1
requests=2.32.3=pypi_0
rich=13.8.0=pypi_0
rsa=4.9=pypi_0
setuptools=73.0.1=pyhd8ed1ab_0
tk=8.6.13=h5083fa2_1
tqdm=4.66.5=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h8827d51_1
uritemplate=4.1.1=pypi_0
urllib3=2.2.2=pypi_0
wcwidth=0.2.13=pypi_0
wheel=0.44.0=pyhd8ed1ab_0
xz=5.2.6=h57fd34a_0
